+++ packages/search/addon/components/navi-search-bar.hbs
