--- packages/search/addon/components/navi-search-result.hbs
@@ -1,1 +1,0 @@
-{{yield}}
