--- packages/search/addon/components/navi-search-result.js
@@ -1,3 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class NaviSearchResultComponent extends Component {}
