--- packages/search/app/components/navi-search-result.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-search/components/navi-search-result';
