+++ packages/search/app/components/navi-search-result/report.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-search/components/navi-search-result/report';
