+++ packages/search/app/services/navi-report-search-provider.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-search/services/navi-report-search-provider';
