+++ packages/search/app/services/navi-search-result.js
@@ -1,0 +1,1 @@
+export { default } from 'navi-search/services/navi-search-result';
