--- packages/search/app/services/navi-search/navi-report-search-provider.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-search/services/navi-search/navi-report-search-provider';
