+++ packages/search/tests/integration/components/navi-search-result/report.js
@@ -3,24 +3,15 @@ import { setupRenderingTest } from 'ember-qunit';
 import { render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 
-module('Integration | Component | navi-search-result', function(hooks) {
+module('Integration | Component | navi-report-search-result', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
     // Set any properties with this.set('myProperty', 'value');
     // Handle any actions with this.set('myAction', function(val) { ... });
 
-    await render(hbs`<NaviSearchResult />`);
+    await render(hbs`<NaviSearchResult::Report />`);
 
     assert.equal(this.element.textContent.trim(), '');
-
-    // Template block usage:
-    await render(hbs`
-      <NaviSearchResult>
-        template block text
-      </NaviSearchResult>
-    `);
-
-    assert.equal(this.element.textContent.trim(), 'template block text');
   });
 });
